Internal FLASH Example
-Test Setup
Single bank flash is used to erase,program and modify the programmed
data.
Data=0x1 is programmed first for 256 bytes starting from 3 different
addresses.
the addresses are for starting of the sector, middle of the sector and
the last 256 bytes of the sector.
After 100 bytes of the 256 byte data 16 bytes are changed to Data=0x2.
-Observation
whole sector is erassed first, data is programmed on the 3 different
addresses.
Data is read and stored in a different buffer to modify it and then the
whole sector gets erassed again to write the modified data in the flash
again.